+++ lit_tests/kernel/wave/gemm.py
@@ -124,70 +124,6 @@ def test_gemm():
 
 
 @run_test
-def test_gemm_global_to_shared():
-    constraints: list[tkw.Constraint] = [tkw.WorkgroupConstraint(M, BLOCK_M, 0)]
-    constraints += [tkw.WorkgroupConstraint(N, BLOCK_N, 1)]
-    constraints += [tkw.TilingConstraint(K, BLOCK_K)]
-    constraints += [tkw.WaveConstraint(M, BLOCK_M / 2)]
-    constraints += [tkw.WaveConstraint(N, BLOCK_N / 2)]
-
-    constraints += [
-        tkw.HardwareConstraint(
-            threads_per_wave=64,
-            mma_type=tkw.MMAType.F32_16x16x16_F16,
-        )
-    ]
-
-    @tkw.wave(constraints)
-    def gemm(
-        a: tkl.Memory[M, K, ADDRESS_SPACE, tkl.f16],
-        b: tkl.Memory[N, K, ADDRESS_SPACE, tkl.f16],
-        c: tkl.Memory[M, N, ADDRESS_SPACE_0, tkl.f32],
-    ):
-        c_reg = tkl.Register[M, N, tkl.f32](0.0)
-
-        @tkw.iterate(K, init_args=[c_reg])
-        def repeat(acc: tkl.Register[M, N, tkl.f32]) -> tkl.Register[M, N, tkl.f32]:
-            a_reg = tkw.read(a)
-            b_reg = tkw.read(b)
-            acc = tkw.mma(a_reg, b_reg, acc)
-            return acc
-
-        tkw.write(repeat, c)
-
-    options = WaveCompileOptions(
-        subs={
-            M: 64,
-            N: 128,
-            K: 64,
-            BLOCK_M: 32,
-            BLOCK_N: 32,
-            BLOCK_K: 16,
-            ADDRESS_SPACE: SHARED_ADDRESS_SPACE,
-            ADDRESS_SPACE_0: GLOBAL_ADDRESS_SPACE,
-        },
-        canonicalize=True,
-        compile_to_mlir=True,
-        use_global_to_shared=True,
-        target="gfx950",
-    )
-    gemm = wave_compile(options, gemm)
-    print(gemm.asm)
-
-    # CHECK-LABEL:    test_gemm
-    # CHECK:          func.func @gemm
-    # CHECK-COUNT-1:    memref.alloc()
-    # CHECK:            scf.for
-    # CHECK:              amdgpu.lds_barrier
-    # CHECK-COUNT-2       amdgpu.gather_to_lds {{.*}} vector<2xf16>
-    # CHECK:              rocdl.s.waitcnt
-    # CHECK:              amdgpu.lds_barrier
-    # CHECK-COUNT-2:      vector.load
-    # CHECK:              amdgpu.mfma
-    # CHECK-COUNT-4:    vector.store
-
-
-@run_test
 def test_non_transposed_gemm():
     constraints: list[tkw.Constraint] = [tkw.WorkgroupConstraint(M, BLOCK_M, 0)]
     constraints += [tkw.WorkgroupConstraint(N, BLOCK_N, 1)]
