+++ tests/kernel/wave/wave_gemm_test.py
@@ -172,12 +172,14 @@ def testPureGemm(
 _xfail = lambda *a: pytest.param(*a, marks=pytest.mark.xfail)
 
 
-_global_to_lds_shapes = [(17, 23, 32), (15, 13, 4), (15, 13, 5)]
+_global_to_lds_shapes = [(17, 23, 32), (15, 13, 4)]
 
 
 @require_e2e
 @require_cdna_3_or_4
-@pytest.mark.parametrize("shape", _global_to_lds_shapes + get_test_shapes("test_gemm"))
+@pytest.mark.parametrize(
+    "shape", _global_to_lds_shapes + get_test_shapes("test_gemm")[:1]
+)
 @pytest.mark.parametrize(
     "enable_scheduling",
     [
@@ -235,10 +237,7 @@ def testGemmGatherToLDS(
     c = device_zeros(shape[0], shape[1], dtype=torch.float32)
     asm = gemm(a, b, c)
 
-    if (shape[2] * datatype.itemsize) % 4 == 0:
-        assert "amdgpu.gather_to_lds" in asm, "gather_to_lds not found in asm"
-    else:
-        assert "amdgpu.gather_to_lds" not in asm, "gather_to_lds found in asm"
+    assert "amdgpu.gather_to_lds" in asm, "gather_to_lds not found in asm"
 
     if run_bench:
         options.benchmark_results_file = perf_filename_iree
